Added velocity feedback from roboteq motor controller Updated launch files with aproximate parameters fixes #43 aero_driver node listens for twist messages and publishes twists with odom twist values fixes #41

+++ roboteq_driver/include/roboteq_driver/roboteq_motor_controller.h
@@ -112,7 +112,7 @@ class RoboteqMotorController{
   /**
    * get the velocity of each motor
    */
-  void getVelocity(uint8_t chan, int32_t& value);
+  void getVelocity(uint8_t chan, double& value);
  private:
   const double maxRPM1_;
   const double maxRPM2_;
